Split view test case file into smaller chunks …and consolidate base test cases and mixins into 1 abstract class for security search views
